refactor: moved shortcuts and onKeyEvents to its own file
